@@ -2,11 +2,11 @@ name: guardian_network
 description: A Guardian Network client
 publish_to: 'none'
 
-version: 0.1.4+3
+version: 0.1.7+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: ^2.10.1
+  sdk: ">=2.16.1 <3.0.0"
+  flutter: ^2.10.3
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_launcher_icons: ^0.9.2
-  flutter_native_splash: ^2.0.4
+  flutter_native_splash: ^2.1.0
   flutter_secure_storage: ^5.0.2
   flutter_svg: ^1.0.3
   google_fonts: ^2.3.1
@@ -23,12 +23,15 @@ dependencies:
   qr_flutter: ^4.0.0
   qr_code_scanner: ^0.7.0
   package_info_plus: ^1.3.1
+  device_info_plus: ^3.2.2
   flutter_sodium: ^0.2.0
-  p2plib: 
+  ntcdcrypto: ^0.2.0
+  cbor: ^5.0.1
+  p2plib:
+    # path: ../p2plib-dart
     git:
       url: git@gitlab.com:guardian-network/social-recovery-p2p/p2plib-dart.git
-      ref: main
-    # git: https://gitlab.com/guardian-network/social-recovery-p2p/p2plib-dart
+      ref: legacy
 
 dev_dependencies:
   flutter_test:
