@@ -2,4 +2,8 @@ import 'package:event_bus/event_bus.dart' as event_bus;
 
 class EventBus extends event_bus.EventBus {}
 
-class RecoveryGroupClearEvent {}
+class RecoveryGroupClearCommand {}
+
+class GuardianShardsClearCommand {}
+
+class GuardianPeersClearCommand {}
