@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:qr_code_scanner/qr_code_scanner.dart';
 
-import '../../../core/utils.dart';
+// import '../../../core/utils.dart';
 import '../../../core/theme_data.dart';
 import '../../../core/widgets/common.dart';
 
@@ -18,7 +18,6 @@ class ScanQRCodePage extends StatefulWidget {
 
 class _ScanQRCodePageState extends State<ScanQRCodePage> {
   final GlobalKey qrKey = GlobalKey(debugLabel: 'QR');
-  Barcode? result;
   QRViewController? controller;
 
   // In order to get hot reload to work we need to pause the camera if the platform
@@ -51,7 +50,6 @@ class _ScanQRCodePageState extends State<ScanQRCodePage> {
 
   @override
   Widget build(BuildContext context) {
-    final state = Provider.of<AddGuardianController>(context);
     return Column(
       children: [
         // Header
@@ -71,14 +69,7 @@ class _ScanQRCodePageState extends State<ScanQRCodePage> {
                 onQRViewCreated: _onQRViewCreated,
               ),
             )),
-        Expanded(
-            child: TextButton(
-          child: const Text('Create random code'),
-          onPressed: () {
-            state.guardianCode = getRandomString(64);
-            state.nextScreen();
-          },
-        )),
+        Expanded(child: Container()),
         Padding(
           padding: const EdgeInsets.only(left: 95, right: 95),
           child: Container(
