@@ -1,28 +1,37 @@
 import 'package:flutter/material.dart';
+import 'package:p2plib/p2plib.dart' show PubKey;
 
 import '../../core/controller/page_controller.dart';
-import '../recovery_group_model.dart';
+import '../../core/model/p2p_model.dart';
 
 class AddSecretController with ChangeNotifier, PagesController {
-  AddSecretController({
-    required int pagesCount,
-    required this.groupName,
-  }) {
+  AddSecretController(
+      {required int pagesCount,
+      required this.groupName,
+      required this.p2pNetwork}) {
     this.pagesCount = pagesCount;
+    p2pNetwork.listen(
+      (p2pPacket) {
+        if (p2pPacket.status == MessageStatus.success &&
+            p2pPacket.type == MessageType.setShard) {
+          _peers.add(p2pPacket.peerPubKey!);
+          notifyListeners();
+        }
+      },
+      onError: (Object e, StackTrace s) {
+        error = e.toString();
+        stackTrace = s.toString();
+        notifyListeners();
+      },
+    );
   }
 
   final String groupName;
+  final Stream<P2PPacket> p2pNetwork;
+  final Set<PubKey> _peers = {};
   String secret = '';
-  final Map<String, RecoveryGroupGuardianModel> _guardians = {};
+  String? error;
+  String? stackTrace;
 
-  Map<String, RecoveryGroupGuardianModel> get guardians => _guardians;
-
-  Future<void> distributeSecret(
-      Map<String, RecoveryGroupGuardianModel> guardian) async {
-    for (var peer in guardian.values) {
-      await Future.delayed(const Duration(seconds: 1));
-      _guardians[peer.name] = peer;
-      notifyListeners();
-    }
-  }
+  Set<PubKey> get peers => _peers;
 }
